[PL] SD: Extract LocalAsmInterface in own file.
The interface is needed in the process, but not the whole
FEM part.
